Fix some problems with the automated tests
Some problems were introduced to the automated tests with one of the recent PRs. This commit fixes those issues and makes a couple of small improvements to our server through the following changes: Move Game UUID generation into the Game object and move tests for this functionality from the CodeGenerator test file to the existing Game test file; Remove Game UUID generation from the custom-objects file and delete the CodeGenerator test file since it was only testing that functionality; Add missing Game properties to a test in the router tests; Make all calls to 'getNewGif(game)' in router 'await' the result before moving on; Condense the response for the 'startgame' action into one message and adjust the Game creation to use the new definition with UUID generation built into the Game itself.