make the Giphy API use a random offset for trending and search
We were getting a lot of repeat results while testing, so we decided to make the Giphy API integration use a random offset on each trending and search request instead of using a static offset that was controlled by the game. This commit updates the appropriate files to remove the static offset and adds an offset generator to the Giphy integration file.