Modify Giphy API to ensure offset randomizer doesn't exceed number of results
Using a custom theme was occasionally returning no results, and after a bit of investigation, it appears that the randomizer was occasionally exceeding the number of results for a custom theme. To avoid this, a Game is now initialized with a gifOffsetMax property that tells the Giphy API what the max offset can be. This property is only used for custom themes (the search endpoint), as the trending endpoint does not have this problem. Tests, mocks, and other relevant files have been updated to reflect this change.